requirements.txt
python-telegram-bot
selenium
beautifulsoup4
requests
webidriver-manager
